+++ lib/xbox_live_api/session_info.rb
@@ -1,5 +1,6 @@
 class XboxLiveApi
   class SessionInfo
+    require 'date'
 
     # @return [Fixnum] the xbox live id for the logged in user
     attr_reader :user_id
@@ -7,11 +8,19 @@ class XboxLiveApi
     attr_reader :gamertag
     # @return [String] the xbox live token used to make requests for the logged in user
     attr_reader :token
+    # @return [DateTime] the time of the token expiration
+    attr_reader :expires
 
-    def initialize(user_id: nil, gamertag: nil, token: nil)
+    def initialize(user_id: nil, gamertag: nil, token: nil, expires: nil)
       @user_id = user_id
       @gamertag = gamertag
       @token = token
+      @expires = DateTime.parse(expires)
+    end
+
+    # @return [Boolean] true if session info will expire within 2 hours
+    def expired?
+      @expires < (DateTime.now + 2/24.0)
     end
 
     def ==(o)